+++ examples/volumetric/tet_cutMesh1.py
@@ -1,6 +1,6 @@
 """Cut a TetMesh with
-an arbitrary polygonal Mesh
-Units are \mum"""
+an arbitrary polygonal Mesh.
+Units are :mum."""
 from vedo import *
 
 settings.use_depth_peeling = True
@@ -11,4 +11,4 @@ sphere = Sphere(r=500, c='g').x(400).alpha(0.2)
 
 tetmesh.cut_with_mesh(sphere, invert=True)
 
-show(tetmesh, sphere, __doc__, axes=dict(xtitle='x [\mum]')).close()
+show(tetmesh, sphere, __doc__, axes=dict(xtitle='x [:mum]')).close()