+++ docs/changes.md
@@ -14,6 +14,8 @@
 - add `settings.backend_autoclose` (default is True) to automatically close `Plotter` after show 
 in jupyter notebooks, #802
 - fix k3d backend - remove pinning to 2.7.4 #808
+- in `Text2d` and `Text3D` shortcut to symbols are now formatted as `:gamma` and no more as `\gamma`
+
 
 
 -------------------------